Rename summarization files
